Fix Compilation Error (`uint32_t` -> `size_t`)

+++ tests/fuzzer.c
@@ -2153,8 +2153,8 @@ static int basicUnitTests(U32 const seed, double compressibility)
     DISPLAYLEVEL(3, "test%3i : table cleanliness through index reduction : ", testNb++);
     {
         int cLevel;
-        uint32_t approxIndex = 0;
-        uint32_t maxIndex = ((3U << 29) + (1U << ZSTD_WINDOWLOG_MAX)); /* ZSTD_CURRENT_MAX from zstd_compress_internal.h */
+        size_t approxIndex = 0;
+        size_t maxIndex = ((3U << 29) + (1U << ZSTD_WINDOWLOG_MAX)); /* ZSTD_CURRENT_MAX from zstd_compress_internal.h */
 
         /* vastly overprovision space in a static context so that we can do all
          * this without ever reallocating, which would reset the indices */
